+++ server/src/services/message_handler.rs
@@ -205,7 +205,7 @@ impl MessageHandler {
                 if let Err(response) = self.validate_request(&mut session, &request.request_header) {
                     Some(response)
                 } else {
-                    Some(self.method_service.call(&mut session, request)?)
+                    Some(self.method_service.call(&address_space, request)?)
                 }
             }
             _ => {